# apps/demo-web/backend/openapi.json
{
  "components": {
    "schemas": {
      "EventAccepted": {
        "properties": {
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "duplicate": {
            "default": false,
            "title": "Duplicate",
            "type": "boolean"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "event_timestamp": {
            "format": "date-time",
            "title": "Event Timestamp",
            "type": "string"
          },
          "status": {
            "const": "accepted",
            "default": "accepted",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "correlation_id",
          "event_timestamp"
        ],
        "title": "EventAccepted",
        "type": "object"
      },
      "EventRequest": {
        "properties": {
          "action": {
            "enum": [
              "view",
              "cart",
              "purchase"
            ],
            "title": "Action",
            "type": "string"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 8,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "impression_id": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impression Id"
          },
          "product_id": {
            "minimum": 1.0,
            "title": "Product Id",
            "type": "integer"
          },
          "quantity": {
            "default": 1,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Quantity",
            "type": "integer"
          },
          "request_id": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "session_id": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Session Id",
            "type": "string"
          },
          "user_id": {
            "minimum": 1.0,
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "product_id",
          "action",
          "session_id"
        ],
        "title": "EventRequest",
        "type": "object"
      },
      "EventStatus": {
        "properties": {
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "feature_service_available": {
            "default": true,
            "title": "Feature Service Available",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "accepted",
              "feature_store_updated"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "status"
        ],
        "title": "EventStatus",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Product": {
        "properties": {
          "brand_id": {
            "title": "Brand Id",
            "type": "integer"
          },
          "brand_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Name"
          },
          "category_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Code"
          },
          "category_id": {
            "title": "Category Id",
            "type": "integer"
          },
          "current_price": {
            "title": "Current Price",
            "type": "number"
          },
          "price_bucket": {
            "title": "Price Bucket",
            "type": "integer"
          },
          "product_id": {
            "title": "Product Id",
            "type": "integer"
          },
          "product_name": {
            "title": "Product Name",
            "type": "string"
          }
        },
        "required": [
          "product_id",
          "product_name",
          "category_id",
          "brand_id",
          "current_price",
          "price_bucket"
        ],
        "title": "Product",
        "type": "object"
      },
      "ProductPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Product"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ProductPage",
        "type": "object"
      },
      "RecommendationItem": {
        "properties": {
          "impression_id": {
            "title": "Impression Id",
            "type": "string"
          },
          "item_id": {
            "title": "Item Id",
            "type": "integer"
          },
          "product": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Product"
              },
              {
                "type": "null"
              }
            ]
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "item_id",
          "score",
          "impression_id"
        ],
        "title": "RecommendationItem",
        "type": "object"
      },
      "RecommendationRequest": {
        "properties": {
          "session_id": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "top_k": {
            "default": 10,
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          },
          "user_id": {
            "minimum": 1.0,
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "RecommendationRequest",
        "type": "object"
      },
      "RecommendationResponse": {
        "properties": {
          "ab_experiment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ab Experiment Id"
          },
          "ab_variant": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ab Variant"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecommendationItem"
            },
            "title": "Items",
            "type": "array"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "request_id",
          "user_id",
          "model_version",
          "items"
        ],
        "title": "RecommendationResponse",
        "type": "object"
      },
      "User": {
        "properties": {
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "segment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segment"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "User",
        "type": "object"
      },
      "UserPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/User"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "UserPage",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "RecSys Demo API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/events": {
      "post": {
        "operationId": "create_event_api_events_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Event"
      }
    },
    "/api/events/{event_id}/status": {
      "get": {
        "operationId": "event_status_api_events__event_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Event Status"
      }
    },
    "/api/products": {
      "get": {
        "operationId": "products_api_products_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Products"
      }
    },
    "/api/recommendations": {
      "post": {
        "operationId": "recommendations_api_recommendations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recommendations"
      }
    },
    "/api/users": {
      "get": {
        "operationId": "users_api_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Users"
      }
    },
    "/healthz": {
      "get": {
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Healthz Healthz Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Healthz"
      }
    },
    "/metrics": {
      "get": {
        "operationId": "metrics_metrics_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Metrics"
      }
    },
    "/ready": {
      "get": {
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Ready Ready Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Ready"
      }
    }
  }
}